+++ public/assets/logo/mobisec-logo-white.png
